extract base tests into its own file
